Create visualization directory and move visualization scripts
