+++ tests/test_plotting.py
@@ -1,6 +1,5 @@
 import numpy as np
-from ms2deepscore.plotting import (create_confusion_matrix_plot)
-from ms2deepscore.train_new_model.visualize_results import plot_histograms, calculate_histograms
+from ms2deepscore.visualize_results.plotting import (create_confusion_matrix_plot, plot_histograms, calculate_histograms)
 
 mock_reference_scores = np.random.random((100, 100))
 mock_comparison_scores = np.random.random((100, 100))
